before confrim menu

+++ Filter/inc/filters/kalman_1D/kalman_filter.h
@@ -8,22 +8,21 @@
 #ifndef KALMAN_FILTER_H_
 #define KALMAN_FILTER_H_
 
-#include "stdint.h" // To get uint16_t. How come I dont need this in Experimentation2?
-
+//#include "stdint.h" // To get uint16_t. How come I dont need this in Experimentation2?
+#include "arm_math.h"
 
 struct Data1D {
 	// Dynamic variables
-	uint16_t x_est;
-	uint16_t x_pred;
-	uint16_t MSE_pred;
-	// Fixed variables (For now)
-	uint16_t A;
-	uint16_t B;
-	uint16_t H;
-	uint16_t C_w;
-	uint16_t C_u;
+	float32_t x_est;
+	float32_t x_pred;
+	float32_t MSE_pred;
+	float32_t A;
+	float32_t B;
+	float32_t H;
+	float32_t C_w;
+	float32_t C_u;
 };
 
-void kalman_filter( struct Data1D* Data, uint16_t y);
+void kalman_filter( struct Data1D* Data, float32_t y);
 
 #endif /* KALMAN_FILTER_H_ */
